requirements_optimized.txt
# OPTIMIZED REQUIREMENTS FOR SOFI AI - MEMORY EFFICIENT
# ====================================================

# Core Flask (lightweight configuration)
Flask==2.3.3
flask-cors==4.0.0
Flask-Limiter==3.5.0
gunicorn==21.2.0

# Database (optimized versions)
supabase==2.0.0
postgrest==0.10.8

# AI (use lighter models when possible)
openai==1.3.0

# Telegram bot (choose one, not both)
pyTelegramBotAPI==4.14.0

# Image processing (lighter alternatives)
pillow==10.0.1
# Removed: opencv-python (heavy), pytesseract (heavy)

# Essential utilities
python-dotenv==1.0.0
requests==2.31.0

# Data processing (lightweight)
# Removed: numpy, pandas (heavy for production)

# Memory monitoring
psutil==5.9.6

# Security
PyJWT==2.8.0

# Async support (minimal)
asyncio-mqtt==0.16.1

# Testing (dev only - not installed in production)
# pytest==7.4.3
# pytest-mock==3.12.0
